// components/panel/targets/panelTargets.enum.ts
export enum PanelTargets {
  contentBody = 'contentBody',
  contentWrapper = 'contentWrapper',
  header = 'header',
  label = 'label',
  modal = 'modal',
  root = 'root',
  modalBackdrop = 'modalBackdrop',
  modalClose = 'modalClose',
  modalDialog = 'modalDialog',
  modalHeader = 'modalHeader',
  title = 'title',
}